// PickEmLeagueFrontEnd/ClientApp/src/Apis/models/GameResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PickEmLeague
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum GameResult {
    NotPlayed = 'NotPlayed',
    HomeWin = 'HomeWin',
    AwayWin = 'AwayWin',
    Tie = 'Tie'
}

export function GameResultFromJSON(json: any): GameResult {
    return GameResultFromJSONTyped(json, false);
}

export function GameResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameResult {
    return json as GameResult;
}

export function GameResultToJSON(value?: GameResult | null): any {
    return value as any;
}